marketing360/marketing360/requirements.txt
django-admin-thumbnails
django-ckeditor-5